Вносит правки на добавленные модули
