+++ js/main.js
@@ -1,7 +1,7 @@
-import {getPhotosData} from './data.js';
+import { getPhotosData } from './data.js';
 
 const SIMILAR_ARRAY_COUNT = 26;
 
 /*eslint-disable no-unused-vars*/
-const PICTURES_DATA =  getPhotosData(SIMILAR_ARRAY_COUNT);
+const PICTURES_DATA = getPhotosData(SIMILAR_ARRAY_COUNT);
 /* eslint-enable no-unused-vars */
